move autolistener into util (#780)
