+++ util/autoListener.go
@@ -1,4 +1,4 @@
-package agent
+package util
 
 import (
 	"fmt"
